Add files via upload
I’m glad to share that I have successfully completed a real-time research project with the support and guidance of "SAK INFORMATIVES". This project provided me with valuable hands-on experience and deep insights into practical applications of advanced concepts. Grateful for the opportunity to work on a live project environment and enhance my skills through this experience.